server/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
filelock==3.15.4
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.7
psycopg2==2.9.9
psycopg2-binary==2.9.9
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0
setuptools==70.3.0
sqlparse==0.5.0
tldextract==5.1.2
urllib3==2.2.2
zope.event==5.0
zope.interface==6.4.post2
